Noise and Biome values are now perfectly in sync with server
The noise and biome values are some slight deviations that produced value mismatched in the "0.001" error range. This will at time lead to the wrong biome or noise value being attribute to a hex. This has been fixed.

Biome type when derived from server was wrong because the json data was not properly converted. This has been fixed